Fixed the Get Rent Breakdowns query.
The previous query itself would have been fine if not for the horrid database, where unique keys are not actually all that unique, infact they get repeated thousands of times in some cases.

Due to this issue, plus due to there being no direct link between the tenancy references and charges (most of the time), some records got affected in a way that when trying to get charges related to certain tenancy refs, you'd get the charge of tenancy ref in question, plus some outdated charges for ~20'000 other properties. This would make a lot of things crash.

Talking to people from bussiness team, I came up with this. The idea here is to calculate when the charge is supposed to take effect (based on some business rules), and get just that (to make it more specific - to avoid getting all other unrelated charges).

The thing is some charges take into effect at exactly april 1st every year, meanwhile others during first monday of april, which keeps on changing from year to year. Hence, you have 3 cases - all old charges apply prior to april 1st, then only 1st of april charges apply in between april 1st and first monday, and then after 1st monday - both new (financial year) charges apply.

Made use sql variables so it would be easier to follow what's going on. Also fixed SQL injection vulnarability for this endpoint - the SQL parameter syntax was used incorrectly making it useless.

Oh, yes, the 1st mondays calculation - the idea is that 6th day is always either the start of second month week, or it is the end of the first week. With SQL if it's either the new week or the end of the current week - the week difference between week before this one and the week in question is considered to be 1. So when you do this WEEK subtract, add magic, the context of individual days is lost (it's like rounding up a number). So you always end up adding whole weeks to zero, always ending up on monday.